Began implementing Eggs
Started by setting up a system to not start needs timers. next need to mess with the food and play menu options to not be available in egg mode.

need to keep track of this stuff in the future too, saving and loading will be a nightmare if I'm not careful