Set up base game save system
doozy of a system but I think it may just work. It basically checks the whole game for objects that implement a saving class, adds that to a list of references that it will use to get all of the save data it needs. it then compiles that list of data it loaded to a variable that can be accessed by whatever needs it, using a string password. Do need to do something about conflicting passwords but for now I'll probably just track it somewhere else