PET SAVING WORKS
Currently has support for 3 save slots for pets. Next task HAS to be setting up a system for determining what pet to spawn though. Will probably have a system for remembering what pet was active last, then a system for changing with one to use. Also need to make sure I add more points in which the game is saved.