Fixed some menu manager issues
weird issue with pausing. Needs some work to make it not overwrite another menu's paused status. other than that I fixed some weird issues with the save slot menu.

at this point menus are mostly taken care of. Should probably move on to either finishing the rest of the device buttons or updating the minigames