Minimized Mode Complete
When entering minimized mode, the window's size will now auto adjust itself to be as small as possible.

There will probably need to be some more work in the future as I add multiple game scales, but for the time being this is works great. Also need to add setting for choosing a monitor to anchor to.