added a stopgap to pet feeding
There was an issue encountered by a demo playtester where the pet was stuck trying to eat while the food was far away. I do not know exactly what caused it, however, I do know that it would be fixed if I were to create a timer to force a pet from eating for more than a minute.