Created a mood behavior + added new plugin
Added a plugin to make navigating larger scripts a lot easier.

Pets now have a mood behavior function that gets called every frame. The function will work by running a randomized timer that will randomly cause the pet to send out a mood to the thought bubble. Need to finish some of the moods in order to properly test it.